Crate seahorse

The Seahorse cryptographic keystore library

This crate provides a ledger-agnostic toolkit for building cryptographic keystores for implementations of the [jf_cap] protocol.

The main types that users of this crate will interact with are Keystore and KeystoreBackend. The former represents the generic keystore interface that this crate provides, including basic keystore functionality like generating and viewing keys or sending and receiving assets. The latter represents the interface to a specific ledger that the user must provide in order to instantiate the ledger-agnostic Keystore interface for their particular system.

Users should also be familiar with [reef], which provides traits to adapt a particular CAP ledger to the ledger-agnostic interfaces defined here.

Re-exports

pub use crate::asset_library::AssetInfo;

pub use crate::asset_library::MintInfo;

pub use crate::txn_builder::RecordAmount;

pub use jf_cap;

pub use reef;

Modules

accounts

asset_library

Collections of and information on CAP assets.

cli

The generic CAP Keystore frontend

encryption

Symmetric encryption for locally persistent keystore data.

events

Event definitions for ledger state changes.

hd

Hierarchical deterministic key generation

io

I/O interface for the CLI.

loader

Traits and types for creating and loading keystores.

persistence

Ledger-agnostic implementation of [KeystoreStorage].

reader

Interactive input.

txn_builder

Transaction building.

Macros

cli_write

Convenience macro for panicking if output fails.

cli_writeln

Convenience macro for panicking if output fails.

command

Create a Command from a help string and a function.

count

Structs

AssetNotFreezableSnafu

SNAFU context selector for the KeystoreError::AssetNotFreezable variant

AssetNotMintableSnafu

SNAFU context selector for the KeystoreError::AssetNotMintable variant

AssetNotViewableSnafu

SNAFU context selector for the KeystoreError::AssetNotViewable variant

AssetVerificationSnafu

SNAFU context selector for the KeystoreError::AssetVerificationError variant

BadMerkleProofSnafu

SNAFU context selector for the KeystoreError::BadMerkleProof variant

BincodeSnafu

SNAFU context selector for the KeystoreError::BincodeError variant

CancelledSnafu

SNAFU context selector for the KeystoreError::Cancelled variant

CannotDecryptMemoSnafu

SNAFU context selector for the KeystoreError::CannotDecryptMemo variant

ClientConfigSnafu

SNAFU context selector for the KeystoreError::ClientConfigError variant

CryptoSnafu

SNAFU context selector for the KeystoreError::CryptoError variant

EncryptionSnafu

SNAFU context selector for the KeystoreError::EncryptionError variant

FailedSnafu

SNAFU context selector for the KeystoreError::Failed variant

InvalidAddressSnafu

SNAFU context selector for the KeystoreError::InvalidAddress variant

InvalidBlockSnafu

SNAFU context selector for the KeystoreError::InvalidBlock variant

InvalidFreezerKeySnafu

SNAFU context selector for the KeystoreError::InvalidFreezerKey variant

InvalidViewerKeySnafu

SNAFU context selector for the KeystoreError::InvalidViewerKey variant

IoSnafu

SNAFU context selector for the KeystoreError::IoError variant

KeySnafu

SNAFU context selector for the KeystoreError::KeyError variant

KeyStreamState

The number of keys of each type which have been generated.

Keystore

The generic CAP keystore implementation.

KeystoreSession

Transient state derived from the persistent KeystoreState.

KeystoreSharedState

Keystore state which is shared with event handling threads.

KeystoreState

The data that determines a keystore.

MnemonicSnafu

SNAFU context selector for the KeystoreError::MnemonicError variant

NoSuchAccountSnafu

SNAFU context selector for the KeystoreError::NoSuchAccount variant

NullifierAlreadyPublishedSnafu

SNAFU context selector for the KeystoreError::NullifierAlreadyPublished variant

PersistenceSnafu

SNAFU context selector for the KeystoreError::PersistenceError variant

StorageTransaction

Interface for atomic storage transactions.

TimedOutSnafu

SNAFU context selector for the KeystoreError::TimedOut variant

TransactionSnafu

SNAFU context selector for the KeystoreError::TransactionError variant

UndefinedAssetSnafu

SNAFU context selector for the KeystoreError::UndefinedAsset variant

UserKeyExistsSnafu

SNAFU context selector for the KeystoreError::UserKeyExists variant

Enums

KeystoreError

Traits

Captures

KeystoreBackend

The interface required by the keystore from a specific network/ledger implementation.

SendFuture

Functions

new_key_pair