Module seahorse::asset_library

Collections of and information on CAP assets.

This module defines AssetInfo and MintInfo, which store auxiliary information about assets which is useful to wallets but not present in [AssetDefinition]. For example, AssetInfo may include a MintInfo, which contains the secret information needed by the asset creator to mint more of that asset type.

This module also defines an interface for verified asset types, VerifiedAssetLibrary. This is a collection of assets which can be signed by a trusted party, such as an application developer, and distributed to client applications. Applications with verified status can thus be displayed as such, for example by including a badge in a GUI application. Note that this library merely provides the mechanisms and interfaces for creating and consuming verified asset libraries. It does not define any specific libraries or verification keys, as these are application-specific and thus should be defined in clients of this crate.

Structs

AssetInfo

Details about an asset type.

AssetLibrary

Indexable collection of asset types.

Icon

A small icon to display with an asset in a GUI interface.

MintInfo

Information required to mint an asset.

VerifiedAssetLibrary

A library of assets, signed by a trusted party.