Module seahorse::encryption

Symmetric encryption for locally persistent keystore data.

All secret data that the keystore stores on the file system is encrypted using this module. The encryption scheme is an encrypt-then-MAC implementation using ChaCha20 as a stream cipher and SHA3-256 as an HMAC.

The choice of stream cipher merits some discussion. Other stream ciphers which were considered include AES-GCM-SIV and AES-CTR. Briefly, the advantages of each:

AES-GCM-SIV is natively authenticating and would not require the use of a separate MAC function. It is also somewhat resistant to nonce reuse, a common class of security bugs in cryptography implementations.

AES-CTR is widely used and thoroughly analyzed, and common consumer hardware includes instructions designed to accelerate this algorithm.

Now, the reasons why we prefer ChaCha20:

AES-GCM-SIV uses a polynomial hash function which is subject to sharding attacks that reduce the security of the cipher to about 90 bits. This figure is extrapolated from table 1 of https://eprint.iacr.org/2017/708.pdf (page 14). Note that the 90 bits figure is based specifically on chosen plaintext attacks, and potentially underestimates the security, since CPA attacks are milder than other attacks (like key recovery) and are usually easiest to carry out anyways. In particular, the length of the longest encrypted plaintext (k in the table indicates the longest message is ~2^k AES blocks long) significantly reduces the additional amount of work required to carry out the attack, but the keystore only encrypts relatively small amounts of data with any one key. Nevertheless, we can avoid this attack entirely by using a hash-based MAC instead of a polynomial MAC.

AES-CTR has security against a plaintext-distinguishing attack which decreases with the length of the message being encrypted, and so it is recommended to rekey after encrypting a certain number of blocks. Detailed recommendations vary from every 1024 blocks (16kb) to every 2^32 blocks (69gb). The lack of a consistent, standardized recommendation coupled with the performance implications of adopting the more conservative recommendation makes this a less attractive choice than ChaCha20.

The root of AES-CTR’s plaintext distinguishing vulnerability is the fact that it is based on a pseudo-random permutation. ChaCha20 is based on a pseudo-random function which is not invertible, and so it is immune to the particular attack that succeeds against AES-CTR. Moreover, even if ChaCha20’s PRF turned out to be invertible, its security parameters (specifically the 512-bit block size) are much better than AES to begin with, and it would take a message of 2^128 blocks being encrypted with the same key to have a 2^-256 probability of distinguishing plaintexts.

The keystore uses this library to encrypt data at a small granularity. For the log-structured data, each log entry is encrypted separately. Matching the encryption granularity to the structure of the data is advantageous for a few reasons. First, it allows us to recover most of the data if a single entry becomes corrupted on disk. More relevantly for this module, it allows us to use a different key for each encrypted message, which is an easy way to avoid nonce reuse misuse vulnerabilities, and works well with procedural key generation.

This module uses the hd module for procedural key generation. A Cipher wraps an entire hd::KeyTree and uses the tree to deterministically generate a new key for each message it encrypts. This reduces are vulnerability to nonce reuse misuse bugs and increases our defense in depth.

Re-exports

pub use hd::Salt;

Structs

Cipher

An authenticating stream cipher.

CipherText

Encrypted and authenticated data.

Enums

Error

Type Definitions

Nonce

Result