Module seahorse::hd

Hierarchical deterministic key generation

This module provides an interface for procedurally generating a tree of keys. It does not implement the full BIP 32 standard for HD wallets, only the features relevant to our system. For example, BIP 32 provides features for deriving public keys from parent public keys without knowing any private keys, to allow for trustless key rotation. Kkey rotation is less important in CAP systems, which provide unlinkability by default.

The keys generated by this interface are 32-byte pseudo-random values with no internal structure. They have no relationship to elliptic curve points, but they can be used as a random seed for generating elliptic curve keys for use in the wallet. The lack of structure makes these keys useful for other applications too, though, such as encryption and decryption.

The relationship between derived keys and their parents can be described as a tree. A single root KeyTree is generated from some external means (such as entropy, or from a password or mnemonic using a KDF). This root tree can be used to derive sub-trees using KeyTree::derive_sub_tree, and any KeyTree can be used to derive more sub-trees as well as leaf keys. Each derived KeyTree and Key is a commitment to its parent, but it is infeasible to compute the parent key from the child.

KeyTrees are opaque. They cannot be used for anything other than to derive more keys. Only leaf keys can be used as keys in applications, via the Key::as_bytes method. Each key commits to its role: leaf keys and key trees will not collide, even if they are derived from the same parent key with the same id.

Structs

Key

A 32-byte pseudo-random key.

KeyTree

A virtual tree of keys.

Mnemonic

A mnemonic representation.

Secret

A wrapper around a secret which cannot be copied.

Type Definitions

Salt

Salt used when deriving keys from passwords.