Struct seahorse::io::SharedIO

pub struct SharedIO { /* fields omitted */ }

Wrapper around an input stream and an output stream

Both the input and output are trait objects, so any types implementing Read and Write can be used here.

SharedIO also has an internal buffer so that it can implement BufRead.

Implementations

impl SharedIO

pub fn new(
input: impl Read + Send + 'static,
output: impl Write + Send + 'static
) -> Self

Construct a new I/O object with particular Read and Write implementations.

pub fn pipe() -> (Self, PipeWriter, PipeReader)

Create a SharedIO instance using a bidirectional pipe.

Returns an IO instance, plus a pair of pipe ends to communicate with the SharedIO. The PipeWriter can be used to send input to the SharedIO, and the PipeReader to read its output.

pub fn std() -> Self

Create a SharedIO instance that uses the standard I/O streams stdin and stdout.