Module seahorse::loader

Traits and types for creating and loading keystores.

This module defines the KeystoreLoader interface, which allows various implementations as plugins to the persistence layer. It also provides a ew generally useful implementations based on the MnemonicPasswordLogin mechanism:

InteractiveLoader

CreateLoader

LoginLoader

RecoveryLoader

Re-exports

pub use create::CreateLoader;

pub use interactive::InteractiveLoader;

pub use login::LoginLoader;

pub use recovery::RecoveryLoader;

Modules

create

interactive

login

recovery

Structs

MnemonicPasswordLogin

Metadata that supports login with a password and backup/recovery with a mnemonic.

Traits

KeystoreLoader