Trait seahorse::loader::KeystoreLoader

pub trait KeystoreLoader<L: Ledger> {
    type Meta;

    fn location(&self) -> PathBuf;
    fn create(&mut self) -> Result<(Self::Meta, KeyTree), KeystoreError<L>>;
    fn load(
        &mut self, 
        meta: &mut Self::Meta
    ) -> Result<KeyTree, KeystoreError<L>>;
}

Required Associated Types

type Meta

Metadata about a keystore which is always stored unencrypted.

This allows loaders and tools to report some basic information about the keystore and handle login attempts without decrypting.

DO NOT put secrets in here.

Required Methods

fn location(&self) -> PathBuf

The location of the keystore targetted by this loader.

This tells users of the loader, like MetaStore, where to load existing metadata from when calling load.

fn create(&mut self) -> Result<(Self::Meta, KeyTree), KeystoreError<L>>

Create a new keystore.

The caller must ensure that no keystore currently exists at self.location(). The loader will create the metadata for a new keystore and return it, along with the key tree to use when deriving keys for the new keystore. The caller should persist the new metadata in self.location().

fn load(&mut self, meta: &mut Self::Meta) -> Result<KeyTree, KeystoreError<L>>

Load an existing keystore.

The caller must have loaded meta from self.location(). The loader will use meta to authenticate the keystore and derive the key tree, which is used to decrypt the rest of the keystore files and derive new keys.

The loader may change meta. For instance, some loaders allow the owner of a keystore to log in using a mnemonic phrase and reset the password. In this case, meta would be updated with a new password. If the value of meta after this call succeeds is not equal to its value before the call, the caller is responsible for persisting the new value in self.location().

If load fails, it will not change meta.