Trait seahorse::KeystoreStorage

pub trait KeystoreStorage<'a, L: Ledger> {
    fn exists(&self) -> bool;

    fn load<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<KeystoreState<'a, L>, KeystoreError<L>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn store_snapshot<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        state: &'life1 KeystoreState<'a, L>
    ) -> Pin<Box<dyn Future<Output = Result<(), KeystoreError<L>>> + Send + 'async_trait>>
    where
        'a: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn store_asset<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        asset: &'life1 AssetInfo
    ) -> Pin<Box<dyn Future<Output = Result<(), KeystoreError<L>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn store_transaction<'life0, 'async_trait>(
        &'life0 mut self, 
        txn: TransactionHistoryEntry<L>
    ) -> Pin<Box<dyn Future<Output = Result<(), KeystoreError<L>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn transaction_history<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<TransactionHistoryEntry<L>>, KeystoreError<L>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn commit<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn revert<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

The interface required by the keystore from the persistence layer.

See persistence::AtomicKeystoreStorage for an implementation of this interface which works for any CAP ledger.

Most code should not use this interface directly. Instead, use KeystoreBackend::store with StorageTransaction for a safer, RAII-based transactional interface.

The persistent storage needed by the keystore is divided into 3 categories, based on usage patterns and how often they change.

Static data. This is data which is initialized when the keystore is created and never changes.

There is no interface in the KeystoreStorage trait for storing static data. When a new keystore is created, the Keystore will call KeystoreBackend::create, which is responsible for working with the storage layer to persist the keystore’s static data.

Dynamic state. This is data which changes frequently, but grows boundedly or very slowly.

This includes the state of the ledger and the set of records owned by this keystore.

Monotonic data. This is data which grows monotonically and never shrinks.

The monotonic data of a keystore is the set of assets types in the asset library and the transaction history.

The storage layer must provide a transactional interface. Updates to the individual storage categories have no observable affects (that is, their results will not affect the next call to KeystoreStorage::load) until KeystoreStorage::commit succeeds. If there are outstanding changes that have not been committed, KeystoreStorage::revert can be used to roll back the entire state of the persistent store to its state at the most recent commit.

This interface is specified separately from the KeystoreBackend interface to allow the implementation to separate the persistence layer from the network layer that implements the rest of the backend with minimal boilerplate. This allows a ledger-agnostic implementation of the storage layer to work with multiple ledger-specific implementations of the networking layer. It is also useful for mocking the storage layer during testing.

Required methods

fn exists(&self) -> bool

Check if there is already a stored keystore with this key.

fn load<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<KeystoreState<'a, L>, KeystoreError<L>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Load the stored keystore identified by the given key.

This function may assume self.exists().

fn store_snapshot<'life0, 'life1, 'async_trait>(
&'life0 mut self,
state: &'life1 KeystoreState<'a, L>
) -> Pin<Box<dyn Future<Output = Result<(), KeystoreError<L>>> + Send + 'async_trait>> where
'a: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Store a snapshot of the keystore’s dynamic state.

fn store_asset<'life0, 'life1, 'async_trait>(
&'life0 mut self,
asset: &'life1 AssetInfo
) -> Pin<Box<dyn Future<Output = Result<(), KeystoreError<L>>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Append a new asset to the growing asset library.

fn store_transaction<'life0, 'async_trait>(
&'life0 mut self,
txn: TransactionHistoryEntry<L>
) -> Pin<Box<dyn Future<Output = Result<(), KeystoreError<L>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Add a transaction to the transaction history.

fn transaction_history<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<Vec<TransactionHistoryEntry<L>>, KeystoreError<L>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

fn commit<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Commit to outstanding changes.

fn revert<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Roll back the persisted state to the previous commit.