Trait seahorse::WalletBackend

pub trait WalletBackend<'a, L: Ledger>: Send {
    type EventStream: 'a + Stream<Item = (LedgerEvent<L>, EventSource)> + Unpin + Send;
    type Storage: WalletStorage<'a, L> + Send;
    fn storage<'l, 'async_trait>(
        &'l mut self
    ) -> Pin<Box<dyn Future<Output = MutexGuard<'l, Self::Storage>> + Send + 'async_trait>>
    where
        'l: 'async_trait,
        Self: 'async_trait;

    fn key_stream(&self) -> KeyTree;

    fn create<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<WalletState<'a, L>, WalletError<L>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn subscribe<'life0, 'async_trait>(
        &'life0 self, 
        from: EventIndex, 
        to: Option<EventIndex>
    ) -> Pin<Box<dyn Future<Output = Self::EventStream> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn get_public_key<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        address: &'life1 UserAddress
    ) -> Pin<Box<dyn Future<Output = Result<UserPubKey, WalletError<L>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn get_nullifier_proof<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        nullifiers: &'life1 mut NullifierSet<L>, 
        nullifier: Nullifier
    ) -> Pin<Box<dyn Future<Output = Result<(bool, NullifierProof<L>), WalletError<L>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn get_transaction<'life0, 'async_trait>(
        &'life0 self, 
        block_id: u64, 
        txn_id: u64
    ) -> Pin<Box<dyn Future<Output = Result<Transaction<L>, WalletError<L>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn register_user_key<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        pub_key: &'life1 UserKeyPair
    ) -> Pin<Box<dyn Future<Output = Result<(), WalletError<L>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn submit<'life0, 'async_trait>(
        &'life0 mut self, 
        note: Transaction<L>, 
        info: TransactionInfo<L>
    ) -> Pin<Box<dyn Future<Output = Result<(), WalletError<L>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn load<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<WalletState<'a, L>, WalletError<L>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn store<'l, 'async_trait, F, Fut>(
        &'l mut self, 
        update: F
    ) -> Pin<Box<dyn Future<Output = Result<(), WalletError<L>>> + Send + 'async_trait>>
    where
        F: Send + FnOnce(StorageTransaction<'a, 'l, L, Self>) -> Fut,
        Fut: Send + Future<Output = Result<StorageTransaction<'a, 'l, L, Self>, WalletError<L>>>,
        'l: 'async_trait,
        F: 'async_trait,
        Fut: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn finalize<'life0, 'async_trait>(
        &'life0 mut self, 
        _txn: PendingTransaction<L>, 
        _txn_id: Option<(u64, u64)>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        L: 'static,
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }
}

Associated Types

type EventStream: 'a + Stream<Item = (LedgerEvent<L>, EventSource)> + Unpin + Send

type Storage: WalletStorage<'a, L> + Send

Required methods

fn storage<'l, 'async_trait>(
&'l mut self
) -> Pin<Box<dyn Future<Output = MutexGuard<'l, Self::Storage>> + Send + 'async_trait>> where
'l: 'async_trait,
Self: 'async_trait,

Access the persistent storage layer.

The interface is specified this way, with the main storage interface in a separate trait and an accessor function here, to allow implementations of WalletBackend to split the storage layer from the networking layer, since the two concerns are generally separate.

Note that the return type of this function requires the implementation to guard the storage layer with a mutex, even if it is not internally shared between threads. This is meant to allow shared access to the storage layer internal, not require it. A better interface would be to have an associated type type<'l> StorageRef: 'l + Deref<Target = Self::Storage> + DerefMut This could be MutexGuard, RwLockWriteGuard, or just &mut Self::Storage, depending on the needs of the implementation. Maybe we can clean this up if and when GATs stabilize.

fn key_stream(&self) -> KeyTree

fn create<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<WalletState<'a, L>, WalletError<L>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

fn subscribe<'life0, 'async_trait>(
&'life0 self,
from: EventIndex,
to: Option<EventIndex>
) -> Pin<Box<dyn Future<Output = Self::EventStream> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

fn get_public_key<'life0, 'life1, 'async_trait>(
&'life0 self,
address: &'life1 UserAddress
) -> Pin<Box<dyn Future<Output = Result<UserPubKey, WalletError<L>>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

fn get_nullifier_proof<'life0, 'life1, 'async_trait>(
&'life0 self,
nullifiers: &'life1 mut NullifierSet<L>,
nullifier: Nullifier
) -> Pin<Box<dyn Future<Output = Result<(bool, NullifierProof<L>), WalletError<L>>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

fn get_transaction<'life0, 'async_trait>(
&'life0 self,
block_id: u64,
txn_id: u64
) -> Pin<Box<dyn Future<Output = Result<Transaction<L>, WalletError<L>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

fn register_user_key<'life0, 'life1, 'async_trait>(
&'life0 mut self,
pub_key: &'life1 UserKeyPair
) -> Pin<Box<dyn Future<Output = Result<(), WalletError<L>>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

fn submit<'life0, 'async_trait>(
&'life0 mut self,
note: Transaction<L>,
info: TransactionInfo<L>
) -> Pin<Box<dyn Future<Output = Result<(), WalletError<L>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Provided methods

fn load<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<WalletState<'a, L>, WalletError<L>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

fn store<'l, 'async_trait, F, Fut>(
&'l mut self,
update: F
) -> Pin<Box<dyn Future<Output = Result<(), WalletError<L>>> + Send + 'async_trait>> where
F: Send + FnOnce(StorageTransaction<'a, 'l, L, Self>) -> Fut,
Fut: Send + Future<Output = Result<StorageTransaction<'a, 'l, L, Self>, WalletError<L>>>,
'l: 'async_trait,
F: 'async_trait,
Fut: 'async_trait,
Self: 'async_trait,

Make a change to the persisted state using a function describing a transaction.

Example

backend.store(key_pair, |mut t| async move {
    t.store_snapshot(wallet_state).await?;
    // If this store fails, the effects of the previous store will be reverted.
    t.store_auditable_asset(wallet_state, asset).await?;
    // Use `t.backend` to access other backend functions during the transaction. Any
    // failures here will revert all previous stores.
    t.backend.do_something().await?;
    Ok(t)
}).await?;

fn finalize<'life0, 'async_trait>(
&'life0 mut self,
_txn: PendingTransaction<L>,
_txn_id: Option<(u64, u64)>
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
L: 'static,
'life0: 'async_trait,
Self: 'async_trait,

Record a finalized transaction.

If successful, txn_id contains the block ID and index of the committed transaction.