Trait seahorse::WalletBackend

pub trait WalletBackend<'a, L: Ledger>: Send {
    type EventStream: 'a + Stream<Item = (LedgerEvent<L>, EventSource)> + Unpin + Send;
    type Storage: WalletStorage<'a, L> + Send;
    fn storage<'l, 'async_trait>(
        &'l mut self
    ) -> Pin<Box<dyn Future<Output = MutexGuard<'l, Self::Storage>> + Send + 'async_trait>>
    where
        'l: 'async_trait,
        Self: 'async_trait;

    fn key_stream(&self) -> KeyTree;

    fn create<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<WalletState<'a, L>, WalletError<L>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn subscribe<'life0, 'async_trait>(
        &'life0 self, 
        from: EventIndex, 
        to: Option<EventIndex>
    ) -> Pin<Box<dyn Future<Output = Self::EventStream> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn get_public_key<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        address: &'life1 UserAddress
    ) -> Pin<Box<dyn Future<Output = Result<UserPubKey, WalletError<L>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn register_user_key<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        pub_key: &'life1 UserKeyPair
    ) -> Pin<Box<dyn Future<Output = Result<(), WalletError<L>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn get_nullifier_proof<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        nullifiers: &'life1 mut NullifierSet<L>, 
        nullifier: Nullifier
    ) -> Pin<Box<dyn Future<Output = Result<(bool, NullifierProof<L>), WalletError<L>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn submit<'life0, 'async_trait>(
        &'life0 mut self, 
        note: Transaction<L>, 
        info: TransactionInfo<L>
    ) -> Pin<Box<dyn Future<Output = Result<(), WalletError<L>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn load<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<WalletState<'a, L>, WalletError<L>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn store<'l, 'async_trait, F, Fut>(
        &'l mut self, 
        update: F
    ) -> Pin<Box<dyn Future<Output = Result<(), WalletError<L>>> + Send + 'async_trait>>
    where
        F: Send + FnOnce(StorageTransaction<'a, 'l, L, Self>) -> Fut,
        Fut: Send + Future<Output = Result<StorageTransaction<'a, 'l, L, Self>, WalletError<L>>>,
        'l: 'async_trait,
        F: 'async_trait,
        Fut: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn finalize<'life0, 'async_trait>(
        &'life0 mut self, 
        _txn: PendingTransaction<L>, 
        _txn_id: Option<(u64, u64)>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        L: 'static,
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }
}

The interface required by the wallet from a specific network/ledger implementation.

This trait is the adaptor for ledger-specific plugins into the ledger-agnostic Wallet implementation. It provides an interface for the ledger-agnostic wallet to communicate with remote network participants for a particular ledger. Implementing this trait for your specific ledger enables the use of the full generic Wallet interface with your ledger.

Associated Types

type EventStream: 'a + Stream<Item = (LedgerEvent<L>, EventSource)> + Unpin + Send

Notifications about ledger state changes.

The wallet must be able to subscribe to a stream of events from the backend. These events occur whenever the ledger state changes, including new blocks being committed and owner memos being published. The backend may also include events when blocks are rejected. This is not required, but it can help the wallet report failures more quickly (if a rejection is not reported, the rejected transaction will eventually be considered timed out).

Each event yielded by the stream must be tagged with an EventSource indicating where the event came from. Implementations of WalletBackend may provide events from multiple sources (for example, block events from a network query service and memo events from a centralized memo store) or they may aggregate all events into a single stream from a single source.

type Storage: WalletStorage<'a, L> + Send

The storage layer used by this implementation.

Required methods

fn storage<'l, 'async_trait>(
&'l mut self
) -> Pin<Box<dyn Future<Output = MutexGuard<'l, Self::Storage>> + Send + 'async_trait>> where
'l: 'async_trait,
Self: 'async_trait,

Access the persistent storage layer.

The interface is specified this way, with the main storage interface in a separate trait and an accessor function here, to allow implementations of WalletBackend to split the storage layer from the networking layer, since the two concerns are generally separate.

Note that the return type of this function requires the implementation to guard the storage layer with a mutex, even if it is not internally shared between threads. This is meant to allow shared access to the storage layer internally, but not to require it. A better interface would be to have an associated type type<'l> StorageRef: 'l + Deref<Target = Self::Storage> + DerefMut This could be [MutexGuard], [async_std::sync::RwLockWriteGuard], or just &mut Self::Storage, depending on the needs of the implementation. This should be cleaned up if and when GATs stabilize.

fn key_stream(&self) -> KeyTree

Get the HD key tree which the wallet should use to generate keys.

This should be configured when the wallet is created, for example, by deriving a key tree from a mnemonic phrase.

fn create<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<WalletState<'a, L>, WalletError<L>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Create a new wallet.

This method should query the current state of the network (or at least some past state) and create a WalletState consistent with that state. It must also persist the initial state by interacting with the storage layer directly.

fn subscribe<'life0, 'async_trait>(
&'life0 self,
from: EventIndex,
to: Option<EventIndex>
) -> Pin<Box<dyn Future<Output = Self::EventStream> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Subscribe to the asynchronous ledgerevent stream.

fn get_public_key<'life0, 'life1, 'async_trait>(
&'life0 self,
address: &'life1 UserAddress
) -> Pin<Box<dyn Future<Output = Result<UserPubKey, WalletError<L>>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Find the public encryption key associated with the given address.

Users in CAP are identified by addresses, but sending to a user also requires access to their public encryption key, for encrypting owner memos. This information must be published in some way, and accessed by the backend in this method.

fn register_user_key<'life0, 'life1, 'async_trait>(
&'life0 mut self,
pub_key: &'life1 UserKeyPair
) -> Pin<Box<dyn Future<Output = Result<(), WalletError<L>>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Publish an encryption key so that other wallets can access it using WalletBackend::get_public_key.

fn get_nullifier_proof<'life0, 'life1, 'async_trait>(
&'life0 self,
nullifiers: &'life1 mut NullifierSet<L>,
nullifier: Nullifier
) -> Pin<Box<dyn Future<Output = Result<(bool, NullifierProof<L>), WalletError<L>>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Determine whether a nullifier is spent and obtain a proof.

nullifiers is a local cache of nullifiers. It may not contain a status or proof for every relevant nullifier. This function should check if nullifier is represented in nullifiers and query the network only if it is not. Optionally, this function may add the proof to the cache after obtaining a proof from the network.

fn submit<'life0, 'async_trait>(
&'life0 mut self,
note: Transaction<L>,
info: TransactionInfo<L>
) -> Pin<Box<dyn Future<Output = Result<(), WalletError<L>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Submit a transaction to a validator.

Provided methods

fn load<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<WalletState<'a, L>, WalletError<L>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Load a wallet or create a new one.

If a wallet exists in storage, it is loaded. Otherwise, WalletBackend::create is called to create a new wallet.

fn store<'l, 'async_trait, F, Fut>(
&'l mut self,
update: F
) -> Pin<Box<dyn Future<Output = Result<(), WalletError<L>>> + Send + 'async_trait>> where
F: Send + FnOnce(StorageTransaction<'a, 'l, L, Self>) -> Fut,
Fut: Send + Future<Output = Result<StorageTransaction<'a, 'l, L, Self>, WalletError<L>>>,
'l: 'async_trait,
F: 'async_trait,
Fut: 'async_trait,
Self: 'async_trait,

Make a change to the persisted state using a function describing a transaction.

Example

backend.store(key_pair, |mut t| async move {
    t.store_snapshot(wallet_state).await?;
    // If this store fails, the effects of the previous store will be reverted.
    t.store_asset(wallet_state, asset).await?;
    // Use `t.backend` to access other backend functions during the transaction. Any
    // failures here will revert all previous stores.
    t.backend.do_something().await?;
    Ok(t)
}).await?;

fn finalize<'life0, 'async_trait>(
&'life0 mut self,
_txn: PendingTransaction<L>,
_txn_id: Option<(u64, u64)>
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
L: 'static,
'life0: 'async_trait,
Self: 'async_trait,

Record a finalized transaction.

If successful, txn_id contains the block ID and index of the committed transaction.

This function is optional and does nothing by default. The backend can override it to perform cleanup or post-processing on completed transactions.